refactor(web): one sidebar ledger — articles, then people
The aside was three differently-styled widgets: a people strip, the
story's own link, and a selected-sources pick. Now it is one list with
two groups — Articles, then People — sharing the same label, gutter,
gaps and hover. The story link is redundant with the first tab, and the
sources tab is the way into the records.

The people tab is renamed Cast: it counts the whole involved cast,
organisations and machines included, while the sidebar's People group
is persons only — the same word on both made the counts contradict.
