+++ apps/web/src/components/stories/Asides/Asides.tsx
@@ -1,14 +1,12 @@
 'use client';
 
-import type {
-	ResolvedContentsEntry,
-	SourceType,
-} from '@mels-loop/content-loaders/types';
+import type { ResolvedContentsEntry } from '@mels-loop/content-loaders/types';
+import { Avatar } from '@mels-loop/ui/primitives';
 import cn from 'classnames';
+import Image from 'next/image';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 
-import { SOURCE_TYPE_ICONS } from '../../../content/sources/source-types';
 import styles from './Asides.module.css';
 
 interface Row {
@@ -19,30 +17,27 @@ interface Row {
 	author?: string;
 }
 
-/**
- * A short editor's pick of sources, closing with a way through to the rest.
- *
- * The full set runs to dozens and lives in its own tab; what belongs beside the
- * prose is a handful worth opening. Rendered as another group in this list
- * rather than as a component of its own, so it inherits the same alignment,
- * hover and active mark as everything above it.
- */
-interface SourcesGroup {
+export interface StoryPerson {
+	href: string;
+	name: string;
+	/** The story-scoped alias ("The Big Boss") or the localized role. */
+	subtitle?: string;
+	avatarUrl?: string;
+}
+
+interface PeopleGroup {
 	label: string;
-	rows: (Row & { type: SourceType })[];
-	moreHref: string;
-	moreLabel: string;
+	people: StoryPerson[];
 }
 
 interface AsidesProps {
 	contents: ResolvedContentsEntry[];
-	sources?: SourcesGroup;
+	people?: PeopleGroup;
 }
 
 interface Group {
 	key: string;
-	/** Absent for the main text, which names itself. */
-	label?: string;
+	label: string;
 	rows: Row[];
 }
 
@@ -51,31 +46,15 @@ interface Group {
  *
  * Deliberately not a faithful rendering of that tree. The tree describes how
  * the story is filed; this describes what a reader wants beside the prose,
- * which is the main text and the articles written around it. Parts marked
- * `aside: false` are reference material that already has its own tab, and
- * repeating them here says nothing about how they relate to the story.
+ * which is the articles written around it. The main text is not repeated here
+ * — it owns the first tab — and parts marked `aside: false` are reference
+ * material that already has its own tab.
  */
 function toGroups(contents: ResolvedContentsEntry[]): Group[] {
 	const groups: Group[] = [];
 
 	for (const entry of contents) {
-		if (entry.type !== 'part') {
-			groups.push({
-				key: entry.ref,
-				rows: [{ href: entry.href, title: entry.title }],
-			});
-			continue;
-		}
-
-		if (entry.aside === false) continue;
-
-		/* A collapsed part is a single page wearing the part's title — the main
-		 * text itself. It leads, with no group label above it. */
-		if (entry.collapse && entry.href) {
-			groups.push({
-				key: entry.ref,
-				rows: [{ href: entry.href, title: entry.title, author: entry.author }],
-			});
+		if (entry.type !== 'part' || entry.aside === false || entry.collapse) {
 			continue;
 		}
 
@@ -97,30 +76,26 @@ function toGroups(contents: ResolvedContentsEntry[]): Group[] {
 }
 
 /**
- * The story's aside: the main text, and the articles written around it.
+ * The story's aside: its people, and the articles written around it.
  *
- * This was a timeline — a vertical rule with dots down the side. The metaphor
- * claimed a sequence these entries do not have: an article about Mel Kaye's CV
- * does not come after the story in time, it sits alongside it. The rail also
- * broke the one placement that matters on a phone, stacked beneath the prose,
- * where a line descending from the text reads as the text continuing rather
- * than as a list of related reading.
- *
- * Every row aligns on the same text edge whatever its depth, and the current
- * page is marked by a rule in the gutter rather than a filled, padded box — so
- * nothing shifts as the reader moves between pages.
+ * One list component on purpose — the people rows and the article rows share
+ * the same group label, gutter, gaps and hover, so the sidebar reads as one
+ * ledger rather than a stack of differently-styled widgets. Every row aligns
+ * on the same text edge, and the current page is marked by a rule in the
+ * gutter rather than a filled, padded box — so nothing shifts as the reader
+ * moves between pages.
  */
-export function Asides({ contents, sources }: AsidesProps) {
+export function Asides({ contents, people }: AsidesProps) {
 	const pathname = usePathname();
 	const groups = toGroups(contents);
 
-	if (groups.length === 0 && !sources) return null;
+	if (groups.length === 0 && !people) return null;
 
 	return (
 		<nav className={styles.root}>
 			{groups.map((group) => (
 				<div key={group.key} className={styles.group}>
-					{group.label && <p className={styles.groupLabel}>{group.label}</p>}
+					<p className={styles.groupLabel}>{group.label}</p>
 					<ul className={styles.list}>
 						{group.rows.map((row) => {
 							const isActive = pathname.endsWith(row.href);
@@ -146,30 +121,46 @@ export function Asides({ contents, sources }: AsidesProps) {
 				</div>
 			))}
 
-			{sources && sources.rows.length > 0 && (
+			{people && people.people.length > 0 && (
 				<div className={styles.group}>
-					<p className={styles.groupLabel}>{sources.label}</p>
+					<p className={styles.groupLabel}>{people.label}</p>
 					<ul className={styles.list}>
-						{sources.rows.map((row) => {
-							const Icon = SOURCE_TYPE_ICONS[row.type];
-							return (
-								<li key={row.href}>
-									<Link href={row.href} className={styles.link}>
-										<span className={styles.title}>
-											<Icon className={styles.typeIcon} aria-hidden="true" />
-											{row.title}
-										</span>
-										{row.subtitle && (
-											<span className={styles.subtitle}>{row.subtitle}</span>
+						{people.people.map((person) => (
+							<li key={person.href}>
+								<Link
+									href={person.href}
+									className={cn(styles.link, styles.personLink)}
+								>
+									<Avatar
+										size="sm"
+										alt=""
+										fallback={person.name
+											.split(' ')
+											.map((part) => part[0])
+											.slice(0, 2)
+											.join('')}
+										image={
+											person.avatarUrl ? (
+												<Image
+													src={person.avatarUrl}
+													alt=""
+													width={64}
+													height={64}
+													className={styles.avatarImage}
+												/>
+											) : undefined
+										}
+									/>
+									<span className={styles.personText}>
+										<span className={styles.title}>{person.name}</span>
+										{person.subtitle && (
+											<span className={styles.author}>{person.subtitle}</span>
 										)}
-									</Link>
-								</li>
-							);
-						})}
+									</span>
+								</Link>
+							</li>
+						))}
 					</ul>
-					<Link href={sources.moreHref} className={styles.more}>
-						{sources.moreLabel}
-					</Link>
 				</div>
 			)}
 		</nav>
