--- apps/web/src/components/stories/StoryPeople/StoryPeople.module.css
@@ -1,58 +1,0 @@
-.root {
-	display: flex;
-	flex-direction: column;
-	gap: var(--ml-space-sm);
-}
-
-.label {
-	margin: 0;
-	font-size: var(--ml-font-size-xs);
-	font-weight: var(--ml-font-weight-bold);
-	text-transform: uppercase;
-	letter-spacing: 0.04em;
-	color: var(--ml-text-muted-color);
-}
-
-.list {
-	margin: 0;
-	padding: 0;
-	list-style: none;
-	display: flex;
-	flex-direction: column;
-	gap: var(--ml-space-sm);
-}
-
-.row {
-	display: flex;
-	align-items: center;
-	gap: var(--ml-space-sm);
-	text-decoration: none;
-	color: inherit;
-}
-
-.avatarImage {
-	width: 100%;
-	height: 100%;
-	object-fit: cover;
-	display: block;
-}
-
-.text {
-	display: flex;
-	flex-direction: column;
-	min-width: 0;
-}
-
-.name {
-	font-size: var(--ml-font-size-sm);
-	color: var(--ml-text-color);
-}
-
-.row:hover .name {
-	text-decoration: underline;
-}
-
-.subtitle {
-	font-size: var(--ml-font-size-xs);
-	color: var(--ml-text-muted-color);
-}
