+++ apps/web/src/locales/en.json
@@ -21,7 +21,8 @@
 		"moreStories": "More stories",
 		"toggleMenu": "Toggle navigation",
 		"closeMenu": "Close navigation",
-		"people": "People"
+		"people": "People",
+		"cast": "Cast"
 	},
 	"menu": {
 		"articles": "Articles",
