+++ apps/web/src/locales/he.json
@@ -21,7 +21,8 @@
 		"moreStories": "סיפורים נוספים",
 		"toggleMenu": "פתיחת תפריט",
 		"closeMenu": "סגירת תפריט",
-		"people": "אנשים"
+		"people": "אנשים",
+		"cast": "משתתפים"
 	},
 	"menu": {
 		"articles": "מאמרים",
